Clean up GPIO glitches
This change lets users choose whether they want glitches suppressed or
not and ensure that consistent messages get sent. This means that if
you're looking for rising edges, you'll only get messages when the GPIO
reads 1. If a GPIO reads 0, then if suppressing glitches, nothing will
be sent. If not surprising glitches a 1 notification still gets sent
since at some point the GPIO was 1. If looking for both edges, glitches
will similarly be suppressed (or not) and messages are guaranteed to
toggle between 0 and 1 and back.

This removes transitions that confuse users like two 0 notifications in
a row when listening on both edges or a 1 when listening for falling
edges and so on.