Add hardware abstraction layer and restrict rpi pull code
This refactors the GPIO code to support pluggable hardware abstraction
layers (HALs) like Circuits.I2C and Circuits.SPI. Currently the options
are sysfs and a stub HAL suitable for testing on non-Linux platforms.

The second change is to pull out the Raspberry Pi-specific code and
limit it to only /dev/gpiomem. The worry is that mmaping /dev/mem is too
much for code in a NIF. Luckily, it seems like /dev/gpiomem is fairly
commonly available on Raspberry Pis. This should hopefully make the NIF
safer and remove a code path that isn't getting tested much.

Support RPi /dev/gpiomem in NIF