Pass build path to Makefile
When switching between mix targets, the C build products need to be
stored in separate build directories or the wrong ones will be used.
This passes the build directory to the Makefile in the $PREFIX
environment variable. Using $PREFIX is a C convention for specifying the
base install directory.

If this project is being included as a dependency, the top level
`mix.exs` must have `build_embedded: true`. If `build_embedded` is
`false`, then the `priv` directory will be a symlink to a common
directory and this won't work.