improve hal_apply_direction/1 for pins not activated by simply exporting
WHY: not rely on the value of direction after export

Because there are pins that have a value for direction after export,
but do not operate according to that value.
These pins will not be activated unless they are overwritten again,
even if they have the same direction value after export.
Therefore, modified it to write regardless of the value of direction.