Support alternative GPIO backends and allow the default to be changed
This makes it possible to use other GPIO controller implementations and
not just the Linux sysfs GPIO implementation. That means that programs
and libraries using Circuits.GPIO can be made to work on non-Nerves
devices, interact with simulated GPIO devices, and be easier to unit
test.

New backends should implement the `Circuits.GPIO.Backend`. The Linux
GPIO implementation does this and is the default. Backends return
`Circuits.GPIO.Handle` implementations when opened so GPIO calls get
dispatched to the right place without any code changes to user programs.
In fact, this change is mostly invisible to Circuits.GPIO users.

To change the default backend, add the following to the `config.exs`:

config :circuits_gpio, default_backend: MyOwnGPIOBackend

Circuits.GPIO currently won't compile the Linux GPIO backend if you do
this. This means that if you're not running on Nerves or Linux device,
you don't have to worry about Linux bits being compiled.