Add hack to handle BBB GPIO remapping
As of Linux 5.15, the AM335x GPIO banks got reordered and this throws
off the pin numbering. See comments for more details. The gist is that
we hide it from existing programs to avoid the new mapping that's
different from almost every Beaglebone doc on the web. This will make
programs work with the caveat that anyone doing manual debugging of
sysfs will notice it, be confused, and hopefully find this commit.

The future has been to switch to cdev, but that's not happening in the
v1.x branch so this is the workaround to allow Linux 5.15 and later.