Remove initial GPIO message and improve timestamp docs
This switches the notification timestamps from cdev to the ones supplied
by the Linux kernel. These will be the most accurate.

When making the switch, there were some details that made keeping the
initial message that gets sent after the `set_interrupts/3` call tough:

1. While Linux stamps cdev event messages with monotonic time, the
   timestamps were 2 seconds off from `clock_gettime(MONOTONIC_TIME)` on
   one device. This made the initial message's timestamp unusable. I
   never figured this out. The Linux kernel looked like it was doing the
   right thing.
2. cdev supports alternative sources of timestamps, so OS monotonic time
   isn't even required to be used. This isn't supported, but could be.
3. The initial message didn't correspond to a transition, so it doesn't
   even help apps that care about this. GenServer restarts, for example,
   could cause confusion.

On top of all of this, Erlang's monotonic time is adjusted compared to
OS monotonic time so these timestamps can't even be compared to times in
Elixir easily. They can just be compared to each other.

The initial notification is a remnant of sysfs. While this is backwards
incompatible with Circuits.GPIO 1.x, it seems like the only rational way
to go.