Add GPIO.line_info/2; improve info results
This changes the label to just be the line label. This matches better
what users can pass to open and seems way friendlier. The controller
label is passed back in the `:controller` since the gpiochip name is
already in the `gpio_spec` part. This additionally passes back the
`:consumer` field which seems too useful not to return.

This commit updates the diagnostics to show the line info since that's
helpful in debugging.