Cache GPIO enumeration results for faster opens
The GPIO enumeration rarely changes so this caches the results in
`:persistent_term`. To be safe, whenever a GPIO isn't found, the
enumeration is redone.

This optimises for the common cases where a set of GPIOs are opened on
boot and then only re-opened on GenServer restarts. It will also help in
the cases where the user does one-off GPIO sets or reads.

As part of this change, more tests were added to the `gpio_spec`
resolution code which fixed a few missed cases where `gpio_spec`'s that
should have resolved didn't.