Remove Line module and add gpio_info type
This calls the information about a GPIO `gpio_info` and clarifies that
it's nondynamic information. Now that `gpio_info` is just a map, there's
more flexibility on backends to provide more or less information
depending on their needs. The `:location` field is the only mandatory
one since enumerating GPIOs is useless without returning one field.

This cleans up some inconsistencies in the docs due to wording changes
in this area.