Add gpio_status/1 function to get runtime info
Once `gpio_info/1` got limited to static information, there wasn't a way
to get runtime configuration and other info. This lets you do that now
and populates the information with consumer, direction and pull mode. It
could be expanded in the future.

This also fixes an issue were `open/3` didn't handle controller labels
with integer offsets.