Fix race to load NIF
This happens when multiple processes call GPIO.open at the same time.
NIFs can only be loaded once only one process gets an `:ok` return. The
others get an error telling them not to do that. That's harmless so just
try calling the function again.

The included unit test starts 32 processes in a race to open the GPIO.
On my MBP, a whole bunch of the processes would get the reload error
before the fix.

Fixes #178