Fix race when counting pins open for tests
The unit tests check that pins properly get cleaned up, but there was a
race condition. This fixes this by using atomics.

Since keeping open pin counts was intended as a test-only feature and
there's a chance of the use of atomics creating compilation issues,
demote the pin count handling to the stub backend only.