Remove deferred loading of NIF
It was possible to crash the BEAM by loading a completely trimmed down
shared library with multiple processes. This converts the NIF loader
back to the traditional `@on_load` way that's serialized by the code
loader. The benefits of delaying the load time are less now that the
shared library has been tested quite a bit.