Fix the backend information to properly reflect test mode
This fixes old documentation on the stub that is no longer relevant with
Circuits.GPIO 2.0. It also properly returns the `Circuits.GPIO.CDev` backend
options that are in effect so that it's possible to tell when you're
using `test: true` (aka the stub).

Before when using the stub:

```
iex(1)> Circuits.GPIO.backend_info()
%{name: Circuits.GPIO.CDev, pins_open: 0}
```

Now:

```
iex(1)> Circuits.GPIO.backend_info()
%{name: {Circuits.GPIO.CDev, [test: true]}, pins_open: 0}
```