Support async GPIO sets in diagnostics
Asynchronously set GPIOs can be checked too quickly and cause an error.
This happens with gpio-sim for the unit tests. It really shouldn't
happen with real hardware since it seems unlikely for messages to be
reordered on gets and sets between different pins on the same
controller. If this actually happens, then showing a stat on the update
delay may be interesting.

The drawback for this change is that real failures will be slower to
report. Given that the timeout is 11ms and the first failure stops the
test, I don't think anyone would notice.